Enormous overhaul of loopNEST, and Poisson rand gen upgraded to 64-bit unsigned int. -MMS

+++ include/NEST/RandomGen.hh
@@ -28,7 +28,7 @@ class RandomGen {
   double rand_exponential(double half_life, double t_min = -1,
                           double t_max = -1);
   double rand_skewGauss(double xi, double omega, double alpha);
-  int poisson_draw(double mean);
+  uint64_t poisson_draw(double mean);
   int64_t binom_draw(int64_t N0, double prob);
   int integer_range(int min, int max);
   vector<double> VonNeumann(double xMin, double xMax, double yMin, double yMax,
